services/manejador-reportes/requirements.txt
# Core Django (alineado con versiones del curso ISIS-2503)
Django==5.2.5
psycopg2-binary==2.9.10

# Production WSGI server (en lugar de runserver)
gunicorn==23.0.0

# Cache Redis (Sprint 2)
redis==5.2.0
django-redis==5.4.0

# Auth0 / JWT (preparado para Etapa 3 - SEG-01)
PyJWT[crypto]==2.10.1
social-auth-app-django==5.6.0
requests==2.32.3

# RabbitMQ (preparado para Etapa 4 - SEG-02)
pika==1.3.2

# Utilidades
python-dotenv==1.1.1
